+++ src/unified-bus/model/protocol/ub-dcqcn.h
@@ -36,11 +36,11 @@ public:
 
     void OnTpAttached(Ptr<UbTransportChannel> tp) override;
     bool IsCcLimited(uint32_t bytes) override;
-    void OnReceiverDataPacketReceived(uint32_t psn,
+    void OnReceiverDataPacketReceived(uint64_t psn,
                                       uint32_t size,
                                       UbIpBasedNetworkHeader header) override;
-    UbCongestionExtTph OnReceiverPrepareAckCongestionHeader(uint32_t psnStart,
-                                                            uint32_t psnEnd) override;
+    UbCongestionExtTph OnReceiverPrepareAckCongestionHeader(uint64_t psnStart,
+                                                            uint64_t psnEnd) override;
     TpOpcode GetAckOpcode() const override;
     void OnSenderDataPacketSent(uint32_t psn, uint32_t size) override;
     void OnSenderRetransmissionPacketSent(uint32_t psn, uint32_t size) override;
